+++ .solcover.js
@@ -1,3 +1,6 @@
 module.exports = {
-  skipFiles: ['contracts/utility/StringUtils.sol'] 
+  skipFiles: [
+    'contracts/utility/StringUtils.sol',
+    'contracts/token/mocks/ZeroTokenMock.sol'
+  ]
 };
